Refines zone identifiers and handling of zones with the same name
Before this change, `Zone.id()` implied that the provider supported
multiple zones with the same name. This turned out to be false, as for
example both CloudDNS and Designate use ids, but enforce zone names
must be unique.

This change introduces a series of changes to support both determining
how zones are identified, as well if the provider supports zones that
have the same name.

One goal was to not add complexity. The overall approach to this is to
only add a concept while removing one.

`Zone.qualifier()` is added to disambiguate zones with the same name.
This is paid for by deprecating `Zone.idOrName()`.

`Provider.zoneIdentification()` is added to that callers can understand
if zone names can be substituted for ids or if qualifiers are supported.
This is paid for by deprecating `Provider.supportsDuplicateZoneNames()`.

Here are the details of this change:
  * Adds `Zone.qualifier()` in support duplicate zones
  * Deprecates `supportsDuplicateZoneNames()` in favor of `Provider.zoneIdentification()` in `NAME`, `OPAQUE` and `QUALIFIED`
  * Deprecates `Zone.idOrName()` as `Zone.id()` cannot be null
  * Adds `ZoneApi.iterateByName()` to support lookups
  * Adds `-n` parameter to CLI zone list